Add files via upload
Q1. Write a program to find all pairs of an integer array whose sum is equal to a given number?

Q2. Write a program to reverse an array in place? In place means you cannot create a new array. You have to update the original array.

Q3. Write a program to check if two strings are a rotation of each other?

Q4. Write a program to print the first non-repeated character from a string?

Q5. Read about the Tower of Hanoi algorithm. Write a program to implement it.

Q6. Read about infix, prefix, and postfix expressions. Write a program to convert postfix to prefix expression.

Q7. Write a program to convert prefix expression to infix expression.

Q8. Write a program to check if all the brackets are closed in a given code snippet.

Q9. Write a program to reverse a stack.

Q10. Write a program to find the smallest number using a stack.